Dev mohit 2

+++ PTS Dictate/View controllers/TabbarVC/TabbarVC.swift
@@ -12,16 +12,16 @@ class TabbarVC: UITabBarController, UITabBarControllerDelegate {
     
     var thirdTabbarItemImageView: UIImageView!
     var viewControllerToSelect: UIViewController?
-    
-    let tabbarNames : [String] = ["Existing\rDictations","Uploads\rin progress","Record","Settings","Logout"]
-    
+        
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        // Do any additional setup after loading the view.
- 
+        self.delegate = self
+        
         let thirdItemView = self.tabBar.subviews[2]
         self.thirdTabbarItemImageView = thirdItemView.subviews.first as? UIImageView
-        self.thirdTabbarItemImageView.contentMode = .center
+        if thirdTabbarItemImageView != nil {
+            self.thirdTabbarItemImageView.contentMode = .center
+        }
         self.navigationController?.addCustomBottomLine(color: .black, height: 2.0)
         additionalSafeAreaInsets.bottom = UIDevice.current.hasNotch ? 0 : 12
         //First, remove the default top line and background
@@ -33,17 +33,21 @@ class TabbarVC: UITabBarController, UITabBarControllerDelegate {
         self.tabBar.addSubview(lineView)
         lineView.layer.masksToBounds = true
         self.tabBar.sendSubviewToBack(lineView)
-//        self.tabBarController?.tabBar.backgroundColor = .white
+        //        self.tabBarController?.tabBar.backgroundColor = .white
+        
+//        let appearance = UITabBarItem.appearance()
+//        let attributes = [NSAttributedString.Key.font:UIFont(name: "HelveticaNeue-Bold", size: 10)]
+//        appearance.setTitleTextAttributes(attributes as [NSAttributedString.Key : Any], for: .normal)
+//        appearance.setTitleTextAttributes(attributes as [NSAttributedString.Key : Any], for: .selected)
     }
-    
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        delegate = self
+        
     }
     
     override func viewWillLayoutSubviews() {
-      
+        
         // access to list of tab bar items
         if let items = self.tabBar.items {
             for (ind, itemt) in items.enumerated() {
@@ -54,15 +58,11 @@ class TabbarVC: UITabBarController, UITabBarControllerDelegate {
                 itemt.imageInsets = UIEdgeInsets(top: 8, left: 0, bottom: -5, right: 0)
                 let viewTabBar = itemt.value(forKey: "view") as? UIView
                 if viewTabBar?.subviews.count == 2 {
-                    // here customise the label
-                    
                     if let label = viewTabBar?.subviews[1] as? UILabel {
                         label.numberOfLines = 2
                         label.textAlignment = .center
-//                        label.lineBreakMode = .byWordWrapping
                         DispatchQueue.main.asyncAfter(deadline: .now()) {
                             label.frame = CGRect(x: label.frame.minX, y: label.frame.minY, width: ind == 0 ? UIScreen.main.bounds.width/9.8 : 44, height: 25)
-                            label.text = self.tabbarNames[ind]
                             viewTabBar?.layoutIfNeeded()
                             viewTabBar?.setNeedsLayout()
                             viewTabBar?.layoutSubviews()
@@ -72,7 +72,7 @@ class TabbarVC: UITabBarController, UITabBarControllerDelegate {
             }
         }
     }
-
+    
     private var bounceAnimation: CAKeyframeAnimation = {
         let bounceAnimation = CAKeyframeAnimation(keyPath: "transform.scale")
         bounceAnimation.values = [1.0, 1.2, 0.9, 1.1, 1.0]
@@ -93,12 +93,20 @@ class TabbarVC: UITabBarController, UITabBarControllerDelegate {
     override func tabBar(_ tabBar: UITabBar, didSelect item: UITabBarItem) {
         let currentIndex = tabBar.items?.firstIndex(of: item)
         let tabbarItemTag = item.tag
+        print(self.navigationItem.hidesBackButton)
         switch tabbarItemTag {
         case 0:
+            if currentIndex == 3{
+                if let settingsVC = self.viewControllers?[3] as? UINavigationController{
+                    settingsVC.popToRootViewController(animated: false)
+                }
+            }
             break
         case 2:
             if currentIndex != selectedIndex {
-                animate(thirdTabbarItemImageView)
+                if thirdTabbarItemImageView != nil {
+                    animate(thirdTabbarItemImageView)
+                }
             }
         default:
             break
@@ -107,30 +115,29 @@ class TabbarVC: UITabBarController, UITabBarControllerDelegate {
     
     func tabBarController(_ tabBarController: UITabBarController, shouldSelect viewController: UIViewController) -> Bool {
         if viewController.title == "logout"{
-            print("logout")
-            CommonFunctions.showAlert(view: viewController, title: "PTS Dictate", message: "Are you sure you want to Logout?", completion: {
-                (result) in
-                if result{
-                    print("Tapped Yes")
-                    CoreData.shared.deleteProfile()
-                    AppDelegate.sharedInstance().moveToLoginVC()
-                }else{
-                    print("Tapped No")
-                }
-            })
-            return false
+            if isUploading{
+                CommonFunctions.alertMessage(view: viewController, title: "PTS Dictate", msg: "Uploading In progress", btnTitle: "Ok", completion: nil)
+                return false
+            }else{
+                CommonFunctions.showAlert(view: viewController, title: "PTS Dictate", message: "Are you sure you want to Logout?", completion: {
+                    (result) in
+                    if result{
+                        print("Tapped Yes")
+                        CoreData.shared.deleteProfile()
+                        AppDelegate.sharedInstance().moveToLoginVC()
+                    }else{
+                        print("Tapped No")
+                    }
+                })
+                return false
+            }
         }else if viewController.title != "Record" {
             if isRecording{
-                audioRecorder?.pause()
                 CommonFunctions.showAlert(view: viewController, title: "PTS Dictate", message: "Do you want to stop the current Recording ?", completion: {
                     (result) in
                     if result{
                         self.setTabBarHidden(true, animated: false)
                         NotificationCenter.default.post(name: Notification.Name("showBottomBtnView"), object: nil)
-                        print("Tapped Yes")
-                    }else{
-                        audioRecorder?.record()
-                        print("Tapped No")
                     }
                 })
                 return false
@@ -147,13 +154,13 @@ extension UIImage {
         //    let rect = CGRect(x: 0.0, y: 0.0, width: 1.0, height: 0.5)
         UIGraphicsBeginImageContext(rect.size)
         let context = UIGraphicsGetCurrentContext()
-
+        
         context!.setFillColor(color.cgColor)
         context!.fill(rect)
-
+        
         let image = UIGraphicsGetImageFromCurrentImageContext()
         UIGraphicsEndImageContext()
-
+        
         return image!
     }
 }
@@ -163,15 +170,15 @@ extension UITabBarController {
         if (tabBar.isHidden == isHidden) {
             completion?()
         }
-
+        
         if !isHidden {
             tabBar.isHidden = false
         }
-
+        
         let height = tabBar.frame.size.height
         let offsetY = view.frame.height - (isHidden ? 0 : height)
         let duration = (animated ? 0.50 : 0.0)
-
+        
         let frame = CGRect(origin: CGPoint(x: tabBar.frame.minX, y: offsetY), size: tabBar.frame.size)
         UIView.animate(withDuration: duration, animations: {
             self.tabBar.frame = frame
